First attempts
I don't know much about the argparse() so I am learnign that here. I know I have a lot to learn here because I had a plan for processing the argparse() inputs but they would have made the inputs on my program far different to those in the Roadmap.sh Project. So, time to change those.